+++ web/app/+examples/index.ts
@@ -2,6 +2,7 @@ export * from './main';
 export * from './modal';
 export * from './input';
 export * from './info-box';
+export * from './vertical-menu';
 export * from './spinner';
 export * from './modificable-list';
 export * from './footer';